+++ packages/yourloops/pages/profile/profile-page-context.hook.ts
@@ -34,7 +34,7 @@ import { REGEX_BIRTHDATE } from '../../lib/utils'
 import { isEqual, some } from 'lodash'
 import { type ProfileErrors, type ProfileForm } from './models/profile-form.model'
 import { type ProfileFormKey } from './models/enums/profile-form-key.enum'
-import { usePatientContext } from '../../lib/patient/patient.provider'
+import { usePatientsContext } from '../../lib/patient/patients.provider'
 import { HcpProfession } from '../../lib/auth/models/enums/hcp-profession.enum'
 import { type Profile } from '../../lib/auth/models/profile.model'
 import { type Settings } from '../../lib/auth/models/settings.model'
@@ -54,7 +54,7 @@ const useProfilePageContextHook = (): UseProfilePageContextHookReturn => {
   const alert = useAlert()
   const { t, i18n } = useTranslation('yourloops')
   const { user, updateProfile, updatePreferences, updateSettings } = useAuth()
-  const patientHook = usePatientContext()
+  const patientsHook = usePatientsContext()
   const isUserPatient = user.isUserPatient()
   const isUserHcp = user.isUserHcp()
 
@@ -144,7 +144,7 @@ const useProfilePageContextHook = (): UseProfilePageContextHookReturn => {
       }
 
       if (isUserPatient) {
-        patientHook.refresh()
+        patientsHook.refresh()
       }
       alert.success(t('profile-updated'))
     } catch (err) {