+++ packages/yourloops/test/unit/components/patient-list/patient-list.hook.test.ts
@@ -26,7 +26,7 @@
  */
 
 import * as authHookMock from '../../../../lib/auth'
-import * as patientProviderMock from '../../../../lib/patient/patient.provider'
+import * as patientProviderMock from '../../../../lib/patient/patients.provider'
 import * as patientListProviderMock from '../../../../lib/providers/patient-list.provider'
 import { act, renderHook } from '@testing-library/react'
 import { usePatientListHook } from '../../../../components/patient-list/patient-list.hook'
@@ -51,7 +51,7 @@ describe('Patient list hook', () => {
     (authHookMock.useAuth as jest.Mock).mockImplementation(() => ({
       getFlagPatients: getFlagPatientsMock
     }));
-    (patientProviderMock.usePatientContext as jest.Mock).mockImplementation(() => ({
+    (patientProviderMock.usePatientsContext as jest.Mock).mockImplementation(() => ({
       searchPatients: searchPatientsMock
     }));
     (patientListProviderMock.usePatientListContext as jest.Mock).mockImplementation(() => ({